pyproject.toml
[tool.poetry]
name = "fast-api"
version = "0.1.0"
description = "My FastAPI project"
authors = ["Asrorbek Orzikulov <asrorbek.arzikulov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
icecream = "^2.1.3"
psycopg = "^3.1.18"
psycopg-binary = "^3.1.18"
sqlalchemy = "^2.0.29"
email-validator = "^2.1.1"
passlib = "^1.7.4"
bcrypt = "4.0.1"
python-jose = "^3.3.0"
cryptography = "^42.0.5"
python-multipart = "^0.0.9"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
motor = "^3.4.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
